backend/pyproject.toml
[project]
name = "ai-equity-research-backend"
version = "0.1.0"
description = "FastAPI backend for AI Equity Research Agent"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.8",
    "uvicorn==0.34.0",
    "pandas==2.2.3",
    "langgraph==0.2.53",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.1",
    "supabase==2.12.0",
    "psycopg==3.2.5",
    "pgvector==0.3.6",
    "google-generativeai==0.8.4",
    "python-multipart==0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
